Update dev-lib
Updates the Travis CI matrix to test against the latest versions of our dependencies.

Also updates the PHPUnit bootstrap to no longer user a deprecated method.

Updates the WordPoints ruleset to avoid errors saying that it isn't found. This way, the WordPoints ruleset isn't required to be registered with PHPCS.